requirements.txt
click==7.1.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.2
pytesseract==0.3.9
python-dotenv==0.19.2
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
